Rilis v2401.0.0
* Sesuaikan pada versi umum

* Rilis 2307.1.0 (#6927)

* Rilis v2307.1.0

* ok

* ok

* ok

* migrasi

* Update catatan_rilis.md

* ok

* perbaikan validasi data keterangan pada data penduduk (#2583)

* Update Penduduk_model.php

* Fix styling

* Update Penduduk_model.php

* [ci ski] catatan rilis

---------

* perbaikan pengecekan nama desa dengan karakter `/`, `(` dan `)` (#2589)

* [fix] ijinkan karakter / sebagai nama desa

* Fix styling

* perbaiki pesan error

* [fix] tambahkan karakter khusus lain, berlaku juga untuk nama kecamatan

* Fix styling

---------


* perbaikan ambil rilis terbaru dari github jika terdapat batasan harian (#2664)

* [bug] perbaiki release

* perbaiki kondisi

* [ci skip] catatan rilis

---------


---------


* Rilis v2308.0.0 (#6960)

* Rilis v2308.0.0

* perbaiki tema

* Update meta.php

---------


* robots.txt (#7035)

* robots.txt

* menambahkan `<?xml version="1.0" encoding="UTF-8"?>` pada sitemap.xml (#7027)

* Rilis 2308.1.0 (#7061)

* batasi cara menampilkan xss code yang telah tersimpan (#2806)

* pakai library dari https://github.com/esyede/captcha (#2809)

* susulan PR #2711 widget untuk tema esensi


* pembatasan karakter pada input post halaman artikel widget (#2711)

* pembatasan akses unggah file berbahaya melalui RFM (#2671)

* tambahkan validasi ketika edit gambar, hanya perbolehkan file jenis gambar saja

* Fix styling

* Update catatan_rilis.md

---------



* pembatasan karakter pada input post halaman admin inventaris tanah (#2704)

* validasi

* Update Api_inventaris_tanah.php

* Update catatan_rilis.md

---------


* pembatasan karakter pada input post halaman admin pembangunan (#2707)

* filter keamanan

* Fix styling

* [ci skip] catatan rilis

---------


* pembatasan karakter pada input post halaman admin keuangan manual (#2706)

* perbaikan keamanan

* perbaikan celah xss

* [ci skip] catatan rilis

---------


* pembatasan karakter pada input post halaman kirim permohonan surat layanan mandiri web (#2705)

* perbaikan keamanan

* [ci skip] catatan rilis

---------


* pembatasan karakter pada input post halaman informasi publik admin (#2708)


* pembatasan hapus dokumen pada halaman dokumen layanan mandiri web (#2702)

* [security] perbaiki hapus dokumen

* sederhanakan

* [ci skip] catatan rilis

---------


* pembatasan karakter pada input post halaman artikel admin (#2714)

* xss

* Update Web_artikel_model.php

* Fix styling

* [ci skip] catatan rilis

---------


* pembatasan tambah/ubah/hapus file/folder pada RFM sesuai hak akses modul yang diberikan (#2690)

* batasi edit, hapus gambar berdasarkan hak akses user

* batasi edit, hapus gambar berdasarkan hak akses user

* Fix styling

* perbaikan hak akses RFM mengikuti hak akses module

* Fix styling

* [ci skip] catatan rilis

---------


* pembatasan akses file metode LFI (Local File Inclusion) (#2665)

* batasi cara akses file

* Fix styling

* pakai preg_match saja

* Fix styling

* hapus

* perbaiki notifikasi agar tampil diview blade

* Fix styling

* [ci skip] catatan rilis

---------


* peningkatan keamanan pada kerentatan LFI Modul Laporan Penduduk (#2754)

* perbaikan uggah dokumen pada laporan apbdes (#2755)

* pembatasan pengiriman pesan pada layanan mandiri web dengan durasi 60 detik (#2715)

* Update Lapak_model.php

* Update MY_Security.php

* Update ajax_calls.php

* perbaikan security trusted host (#2801)

* perbaikan security trusted host (#2801)


* Rilis 2309.0.0

* composer update

* [ci skip] memutahirkan catatan rilis

* update anjungan controller

* perbaikan susulan

* Update Anjungan.php

* Rilis v2309.1.0

* sesuaikan tampilan

* pembatasan input terhadap serangan xss pada modul pengaduan admin. (#2918)

* default config demo pada desa/config/config.php adalah true yang tidak diaktifkan

* Update Pelanggan_Controller.php

* update .gitignore (#7036)

* penyesuaian token pelanggan

* Rilis v2310.0.0 (#7157)

* rilis v2310.0.0

* run composer update --no-dev

---------

* Rilis v2310.1.0 (#7203)

* Persiapan rilis umum 2310.1.0

* obfuscate

* sesuaikan

* hapus librari data_publik dan Curly

* sesuaikan

* update

* Update opensid_helper.php

* Update Config.php

---------


* cherry-pick commit develbar premium

* Update .gitignore

* Rilis v2311.1.0

* ganti index

* Rilis 2312.0.0 (#7488)

* Rilis 2312.0.0

* Update catatan_rilis.md

---------


* Persiapan rilis umum v2312.1.0 (#7531)

* Rilis v2401.0.0

* perbaikan karena patch keamanan pada artikel

* perbaikan captcha

* [ci skip] memutahirkan catatan rilis

---------




migration timestamp

+++ donjo-app/models/migrations/Migrasi_fitur_premium_2204.php
@@ -308,7 +308,7 @@ class Migrasi_fitur_premium_2204 extends MY_model
             $hasil = $hasil && $this->dbforge->create_table('kehadiran_jam_kerja', true);
         }
 
-        if (Schema::hasTable('kehadiran_jam_kerja') && JamKerja::count() != 7) {
+        if (Schema::hasTable('kehadiran_jam_kerja') && JamKerja::count() != 7 && identitas('id')) {
             if (JamKerja::truncate()) {
                 // Tambahkan data kehadiran jam kerja
                 $hari = [
